Finished restructuring hdf5 base object and locations, still more work to be done.
