+++ src/hdf5objects/__init__.py
@@ -1,1 +1,21 @@
-__version__ = '0.1.0'
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+""" __init__.py
+Description:
+"""
+__author__ = "Anthony Fong"
+__copyright__ = "Copyright 2021, Anthony Fong"
+__credits__ = ["Anthony Fong"]
+__license__ = ""
+__version__ = "1.0.0"
+__maintainer__ = "Anthony Fong"
+__email__ = ""
+__status__ = "Prototype"
+
+# Default Libraries #
+
+# Downloaded Libraries #
+
+# Local Libraries #
+from hdf5object import HDF5Object, HDF5Dataset
+
