+++ src/hdf5objects/datasets/__init__.py
@@ -18,7 +18,3 @@ __status__ = "Prototype"
 
 # Local Libraries #
 
-
-# Definitions #
-# Classes #
-
